# docs/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Campus Analytics API
  description: Backend REST API for Smart Campus Analytics (KPMG Challenge)
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://smart-campus-backend.onrender.com
    description: Production Render service
paths:
...
  /api/v1/students:
    get:
      summary: List students with pagination and filters
      description: Accessible by staff roles (admin, faculty, placement_officer). Filter by department, semester, cohort, status, or search query.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: semester
          schema:
            type: integer
        - in: query
          name: cohort
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, graduated, suspended]
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of student profiles
        '403':
          description: Forbidden - staff role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create student profile
      description: Admin-only. Creates a new student record with unique uppercase studentId.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, firstName, lastName, department, program, semester, enrollmentYear]
              properties:
                studentId:
                  type: string
                  example: "STU_2024_001"
                firstName:
                  type: string
                  example: "Aarav"
                lastName:
                  type: string
                  example: "Sharma"
                department:
                  type: string
                  example: "Computer Science"
                program:
                  type: string
                  example: "B.Tech CSE"
                semester:
                  type: integer
                  example: 6
                enrollmentYear:
                  type: integer
                  example: 2021
      responses:
        '201':
          description: Student profile created successfully
        '409':
          description: Conflict - studentId already exists
...
  /api/v1/segments:
    get:
      summary: List student segments
      description: Staff roles only. Returns segment definitions and aggregate student counts. Member studentIds list is omitted unless includeStudents=true is requested.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: includeStudents
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of defined segments
        '403':
          description: Forbidden - staff role required
...
  /api/v1/intervention-catalog:
    get:
      summary: List intervention catalog entries
      description: Staff roles only. Returns available intervention types, eligibility rules, and capacity units.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: activeOnly
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of intervention catalog entries
        '403':
          description: Forbidden - staff role required

    post:
      summary: Create or update intervention catalog entry
      description: Admin only. Defines eligibility rules, cost/capacity units, and duration.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - interventionType
                - name
                - description
              properties:
                interventionType:
                  type: string
                  example: "remedial_classes"
                name:
                  type: string
                  example: "Remedial Faculty Mentoring"
                description:
                  type: string
                  example: "Weekly mentoring for struggling subjects"
                capacityUnit:
                  type: string
                  default: "seats"
                costUnits:
                  type: number
                  default: 0
                durationDays:
                  type: integer
                  default: 30
                eligibilityRules:
                  type: object
      responses:
        '200':
          description: Catalog entry created or updated
        '403':
          description: Forbidden - admin role required
...
  /api/v1/simulations:
    post:
      summary: Create simulation scenario
      description: Staff roles only. Defines a hypothetical intervention scenario in draft status.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - interventionTypes
              properties:
                strategy:
                  type: string
                  enum: [targeted, uniform, mixed, custom]
                  default: "targeted"
                interventionTypes:
                  type: array
                  items:
                    type: string
                cohortFilters:
                  type: object
                capacityConstraints:
                  type: object
                assumptions:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Scenario definition created in draft status
        '403':
          description: Forbidden - staff role required
...
  /api/v1/interventions:
    get:
      summary: List student interventions
      description: Authenticated users. Staff sees all/filtered interventions; students are automatically scoped to own studentId.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: studentId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [planned, assigned, in_progress, completed, cancelled]
        - in: query
          name: interventionType
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of interventions
...
  /api/v1/feedback:
    post:
      summary: Submit student or faculty feedback
      description: Authenticated users. Students are restricted to submitting feedback for their own studentId.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - feedbackType
              properties:
                studentId:
                  type: string
                feedbackType:
                  type: string
                  enum: [student_satisfaction, faculty_feedback, course_feedback, other]
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 1000
                visibility:
                  type: string
                  enum: [private, staff_only, aggregated]
                  default: "staff_only"
      responses:
        '201':
          description: Feedback submitted successfully
        '400':
          description: Validation error
        '403':
          description: Forbidden - cross-student feedback submission attempt

    get:
      summary: List feedback records
      description: Authenticated users. Students see only their own feedback. Private comments submitted by peers are redacted for non-admin faculty.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: studentId
          schema:
            type: string
        - in: query
          name: feedbackType
          schema:
            type: string
        - in: query
          name: visibility
          schema:
            type: string
      responses:
        '200':
          description: Paginated feedback list
...
  /api/v1/audit/events:
    get:
      summary: List audit trail events
      description: Admin role only. Returns security-relevant administrative audit logs with sanitized credentials.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: resourceType
          schema:
            type: string
        - in: query
          name: actorUserId
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of audit events
        '403':
          description: Forbidden - admin role required
  /api/v1/copilot/query:
    post:
      summary: AI Copilot natural language query assistant
      description: Authenticated users. Grounded query dispatcher routing strictly to verified backend analytics services. Disallows arbitrary SQL/NoSQL injections.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 2
                  maxLength: 500
      responses:
        '200':
          description: Grounded analytical response or explicit not_configured status
        '400':
          description: Validation error or unauthorized query syntax attempt
components:
  schemas:
    Meta:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          example: "req_f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: "email"
        message:
          type: string
          example: "Must be a valid educational email"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT